refactor: 유사한 파일 묶어 폴더로 정리
